TCS: Fix problem with conicSystem minor planets
The DTD defaulted to an attribute called "type" which always returned
a defined value. We were erroneously writing out conicSystem in the
old style using "TYPE". This meant that on read "TYPE" was never
read because "type" always got defined. Now read both and use the
old in preference to new. Also write the new form.