Enhance qsummary output for SCUBA-2
SCUBA-2 has many subtle issues with using a simple target name.
This patch now includes "inbeam" in the output and will look
at the shutter and whether current azimuth is required for the
observation.

Also change "EMPTY" to "TBD" since it really means that we wish
to know later. This distinguishes it from "NONE" state which
means we don't need a target at all and "CurrentAz" which means
we'll look where we happen to be.