Set tracking system based on native coordinate system
Previously the code always chose J2000 if the tracking system
was "TRACKING". Now it looks at the native system that was
used to create the Astro::Coords object. This means that
if you specify GALACTIC in the OT the config will now track
in Galactic and all the map areas and offsets will rotate.