Pong 900 parameters adjusted at high elevations.
See SCUBA-2 Trac ticket #986.

Implemented the fixup method in JAC::OCS::Config so that it
calls a fixup method in JAC::OCS::Config::TCS.
In the event of the observation being an exact 900 x 900
arcsecond CURVY_PONG with either the start or end position
above 70 degrees elevation the DY and VELOCITY parameters are
rewritten.  The number of position angles is adjusted to keep
a similar total time, and the JOS_MIN parameter is recalculated.