Checked percentage change of JOS_MIN in pong fixup.
JAC::OCS::Config::TCS::_fixup_pong_high_el now checks whether the
JOS_MIN parameter increased by more than 5% as a result of adjusting
the scan parameters and number of times round the pong, in which
case it tries to reduce the number of times round the pong.

Also moved the parameters together at the start of the
subroutine and added additional comments.