TCS: specify units for coordinates for spherSystem
When constructing an Astro::Coords object for "spherSystem" targets,
the units should now always be specified.  This is intended to match
the way the TCS would interpret the coordinates: as degrees for
galactic and Az/El coordinates, but as hours and degrees for
B1950, J2000 and HA/Dec coordinates.

This makes use of a new feature of the Astro::Coords constructor methods
whereby a reference to an array of units can be given.

Note that this should prevent guessing between radians and hours/degrees
for J2000 coordinates, for example, but may give unexpected results
for RA as a decimal value.  (From a brief inspection of the TCS code,
it seems that the TCS would interpret a decimal RA value as hours
rather than degrees.)