TCS: Extend check for sexagesimal coordinates
This class used to just look for colons to indicate that coordinates
are sexagesimal.  It now also looks for numbers separated by spaces.
The general problem is that Astro::Coords conflates the units
(e.g. radians or degrees) and base (e.g. decimal or sexagesimal)
into one "units" option, so we have to attempt to guess which
to specify.

See also fault 20191017.003.