XMLHelper: attempt to improve XML indentation
Please see fault 20200522.008.

Some indentation was being increased each time the XML was
rewritten, causing it to grow unnecessarily.  The updated
indentation method still takes the simplistic approach of
pattern-matching against a whole line of XML to try to
determine what is there, but it does now attempt to track
when in the middle of an element, so that attributes on
separate lines can keep their indentation while other lines
have it stripped.  In addition some of the <help> elements
include formulae where relative indentation is important,
so the number of leading spaces being removed is tracked.