Correct handling of rv as redshift
Redshifts were not recognized by JAC::OCS::Config::TCS::BASE
because it was looking for "REDSHIFT" instead of "redshift"
(in lower case as specified by the TCS DTD).  Redshifts were
also not written correctly because JAC::OCS::Config::TCS::Generic
used the velocity accessor ("rv") even if the definition was
"redshift".  (The TCS DTD specifies that when the definition
is "redshift", "rv" should be dimensionless whereas it is a
velocity in km/s otherwise.)  Added a test script to check the
handling of rv in 3 different definitions (including redshift).