TCS: also copy tracking system in setTargetSync
When copying the coordinates to matching tags in setTargetSync,
also set the tracking system to match that in the new base object.
These ("coords" and "tracking_system") are essentially what the
JAC::OCS::Config::TCS::BASE->from_coord method sets up in the new
object.  Currently tracking system is always set to "TRACKING",
to be filled in later when stringifying to XML, but copy the value
in case this changes.

See also fault: 20190628.003